Final Draft
Removed sections in favor of div wrappers because sections need
headings according to the W3C HTML validator.

Changed background color of body and changed input element backgrounds
because the previous shade of grey was hard to look at on top of the
new background.

Added transitions for input element borders.

Removed appearance:none because it was flagged as an error in W3C CSS3
Validator although I left the vendor specific rules because that was
listed as a valid exception for CSS validation errors in the rubric and
I wanted to use my arrow decoration that I drew instead of the default
decoration. This works in Chrome, Safari, and Firefox but might not
work in IE or Edge—I haven’t tested on a PC yet.

Other minor adjustments to layout.